# PI_FrontEnd/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/authenticate": {
            "post": {
                "tags": [
                    "authentication-controller"
                ],
                "operationId": "authenticate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RegisterRequest": {
                "type": "object",
                "properties": {
                    "nom": {
                        "type": "string"
                    },
                    "prenom": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mdp": {
                        "type": "string"
                    },
                    "cin": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "dateNaissance": {
                        "type": "string",
                        "format": "date"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "CLIENT"
                        ]
                    },
                    "numtel": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "adresse": {
                        "type": "string"
                    },
                    "profession": {
                        "type": "string"
                    },
                    "genre": {
                        "type": "string",
                        "enum": [
                            "HOMME",
                            "FEMME"
                        ]
                    },
                    "salaire": {
                        "type": "number",
                        "format": "float"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "banni": {
                        "type": "boolean"
                    }
                }
            },
            "AuthenticationResponse": {
                "type": "object",
                "properties": {
                    "access_token": {
                        "type": "string"
                    },
                    "refresh_token": {
                        "type": "string"
                    }
                }
            },
            "AuthenticationRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "mdp": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}